Add generic_query() to run insert and select over the same function
The tokio-postgres client previously had query() for executing queries
that return rows and execute() for executing queries that do not return
rows.

generic_query() is added to the tokio-postgres client to accept all
queries and return a result that can contain data rows or the number of
rows affected by a query.

The responses are aggregated into a new ResultStream that extracts
DataRow, CommandComplete, and ErrorResponse messsages. The ResultStream
currently does not parse NoticeResponse messages.