Add Row Description Fields to SimpleQuery response
Adds the field information from the Row Description message to
SimpleQueryStream and propagates this information to SimpleQueryRow.
Previously, we took the column name from the fields and stored this as a
SimpleColumn.

This change was designed to limit the impact to the current SimpleQuery
interface. The field information is stored in a new OwnedField struct,
which is identical to the Field struct except that the name is stored as
a String. This allows the SimpleQueryMessage to outlive the
RowDescriptionBody.

We could instead store the RowDescriptionBody and attempt to parse it
when SimpleQueryRow::columns() is called. This is analogous to how the
row body is handled by SimpleQueryRow::get()/try_get(). However, this
approach was not taken to limit the changes to the current interface.

SimpleQueryRow::columns() now requires an allocation to return each
column as a SimpleColumn. This can be avoided by using the new
SimpleQueryRow::fields() method.