Fix row description multi-write
In cases where SELECT would return empty results, an upstream user may
need to still write out a row description. That's why in a previous
commit we updated CommandComplete to include the fields. However, in
cases where rows are being returned, we don't want to accidentally write
more row descriptions than would be appropriate.

This adds logic to ensure that if we have written in data rows that we
don't include the fields in the next follow up command complete we
write. After writing such a command complete, we reset our bool that
keeps track of whether we recently wrote data rows.