Allow specifying result formats for generic_query_raw
The existing frontend::bind function does allow us to specify arbitrary
result formats, but the query::encode_bind function calls it with a
hardcoded result format of Some(1), which causes us to request binary
format for all results.

Since generic_query_raw is already part of our custom patchset that
we've added on top of the upstream tokio-postgres code, I've modified
this function to take result formats directly, and we can change the
ReadySet code that uses this accordingly. Elsewhere, I've tried to avoid
changing any existing APIs so as to make it easier to maintain this
patch set in the future if we decide to merge in changes from upstream.
As such, I've generally erred on the side of adding new functions that
take the extra argument for result formats, and then retain the existing
functions as-is and have them call the new function with the same
hardcoded Some(1) value that retains the prior behavior.