Add upstream PG's CommandCompletion Tag to GenericResult
Currently, the `ResultStream` for a query returns either the rows
that were selected (`GenerciResult::Row`), or the number of rows
modified by the statement (`GenericResult::NumRows`). `NumRows`
is insufficient for Readyset's needs not all statement executed
this way "modify rows". Thus, we need the actual Tag data from
the upstream postgres's CommandCompletion message that was returned
from PG so we can ship it back to our callers. Luckily, the
`Message::CommandComplete` already has this data, we just need to
expose it in the API.

This patch renames `GenericResult::NumRows` to
`GenericResult::Command`, and adds the Tag data into the enum variant.
The important part is adding the tag data to the variant, but the
name change makes the variant more appropriate.

Fixes: https://github.com/readysettech/readyset/issues/1038