Custom patches for readyset tree
This patch is based on tag readyset-20240711.  Previously, we
maintained this repo as a fork and merged with upstream occasionally.
Now we sync to upstream, discard some now unnecessary changes, and
dump all of our custom changes into a single commit so it may be
more easily rebased for as long as this fork may live.