Extend `query_typed_raw()` parameters and return
The `query_typed_raw()` function is the function for executing unnamed
prepared statements. This patch makes two changes:

1) Add `results_format` to the function's parameters. Basically, this
is the same changes we made to support different result formats for
regular, name prepared statements; see [0] and [1]. This allows our
clients to tell us what format it wants, and we can pass it through to
the driver.

2) Alter `RowStream`'s iterator `Item` to be `GenericResult`. The
previous type was `Row`, but that does not allow us to return a
Command. The assumption the driver makes is that only DML statements
will execute as (named or unnamed) prepared statements. However, mahy
other driver, notably pgjdbc, executes SQL commands (like BEGIN and
COMMIT) as unnamed prepared statements. Thus, this patch broadens the
`RowStream` iterator type. We did basically the same thing to support
regular prepared statements; see [2] and [3].

[0] https://github.com/readysettech/readyset/issues/268
[1] https://github.com/readysettech/rust-postgres/pull/6
[2] https://github.com/readysettech/readyset/issues/1038
[3] https://github.com/readysettech/rust-postgres/pull/13